Basic Model tests for most basic types.

+++ Home.TodoistClient/Models/Label.cs
@@ -9,7 +9,7 @@
         /// The id of the label.
         /// </summary>
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public uint Id { get; set; }
 
         /// <summary>
         /// The name of the label.
@@ -23,13 +23,13 @@
         /// And for the additional colors of the premium users: #008299, #03b3b2, #ac193d, #82ba00, #111111.
         /// </summary>
         [JsonProperty("color")]
-        public long Color { get; set; }
+        public uint Color { get; set; }
 
         /// <summary>
         /// Label’s order in the label list (a number, where the smallest value should place the label at the top).
         /// </summary>
         [JsonProperty("item_order")]
-        public long ItemOrder { get; set; }
+        public uint ItemOrder { get; set; }
 
         /// <summary>
         /// Whether the label is marked as deleted (where 1 is true and 0 is false), converted to bool on client.
